+++ tests/test_config_middleware.py
@@ -1,7 +1,6 @@
 from montague.compat.config import ConfigMiddleware
 import pytest
 from webtest import TestApp as WebtestApp
-import six
 
 
 class Bug(Exception):
@@ -10,13 +9,12 @@ class Bug(Exception):
 
 def app_with_exception(environ, start_response):
     def cont():
-        yield "something"
+        yield b"something"
         raise Bug
-    start_response('200 OK', [('Content-type', 'text/html')])
+    start_response(b'200 OK', [(b'Content-type', b'text/html')])
     return cont()
 
 
-@pytest.mark.skipif(six.PY3, reason="ConfigMiddleware requires Paste, which is py2-only")
 def test_error():
     wrapped = ConfigMiddleware(app_with_exception, {'test': 1})
     test_app = WebtestApp(wrapped)